fix(web-ui): rename file contain error message to updateUserMe
